+++ services/blockchain-connector/src/sdk_v5/peers.js
@@ -1,6 +1,6 @@
 /*
  * LiskHQ/lisk-service
- * Copyright © 2021 Lisk Foundation
+ * Copyright © 2020 Lisk Foundation
  *
  * See the LICENSE file at the top-level directory of this distribution
  * for licensing information.
@@ -13,39 +13,102 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const actions = require('./actions');
+const peerCache = require('./peerCache');
+const GeoService = require('../geolocation');
 
-const peerStates = {
-	CONNECTED: 'connected',
-	DISCONNECTED: 'disconnected',
+const addLocation = async (ipaddress) => {
+	try {
+		const result = await GeoService.requestData(ipaddress);
+		return result;
+	} catch (e) {
+		return {};
+	}
 };
 
-const refactorPeer = (orgPeer, state) => {
-	const { ipAddress, options: { height } = {}, ...peer } = orgPeer;
-	peer.state = state;
-	peer.height = height;
-	peer.ip = ipAddress;
-	return peer;
+const getPeers = async params => {
+	let peers;
+
+	const state = params.state ? params.state.toString().toLowerCase() : undefined;
+
+	if (['connected', 'disconnected'].includes(state)) peers = await peerCache.get(state);
+	else peers = await peerCache.get();
+
+	const intersect = (a, b) => {
+		const setB = new Set(b);
+		return [...new Set(a)].filter(x => setB.has(x));
+	};
+
+	const filterParams = ['ip', 'httpPort', 'wsPort', 'os', 'version', 'networkVersion', 'height', 'broadhash'];
+	const activeParams = Object.keys(params).filter(item => params[item]);
+	const activeFilters = intersect(filterParams, activeParams);
+
+	const filteredPeers = peers.filter(peer => {
+		let result = true;
+
+		activeFilters.forEach(property => {
+			if (params[property] !== peer[property]) result = false;
+		});
+
+		return result;
+	});
+
+	const sortBy = (array, p) => {
+		const [property, direction] = p.split(':');
+		if (property === 'version') array.sort((a, b) => a[property] > b[property]);
+		if (property === 'height') array.sort((a, b) => Number(a[property]) < Number(b[property]));
+		if (direction === 'asc') array.reverse();
+		return array;
+	};
+
+	if (params.sort && /^.+:(asc|desc)$/.test(params.sort)) sortBy(filteredPeers, params.sort);
+
+	let sortedPeers = filteredPeers;
+
+	if (params.offset || params.limit) {
+		if (!params.offset) params.offset = 0;
+		sortedPeers = filteredPeers.slice(params.offset,
+			(params.limit || filteredPeers.length) + params.offset);
+	}
+
+	const dataWithLocation = await Promise.all(sortedPeers.map(async (elem) => {
+		elem.location = await addLocation(elem.ip);
+		return elem;
+	}));
+
+	const meta = {
+		count: dataWithLocation.length,
+		offset: params.offset,
+		total: peers.length,
+	};
+
+	return {
+		data: dataWithLocation,
+		meta,
+	};
 };
 
-const getPeers = async () => {
-	const connectedPeers = await actions.getConnectedPeers();
-	connectedPeers.data = connectedPeers
-		.map(orgPeer => refactorPeer(orgPeer, peerStates.CONNECTED));
+const getConnectedPeers = async params => {
+	const response = await getPeers(Object.assign(params, { state: 'connected' }));
+	return response;
+};
 
-	const disconnectedPeers = await actions.getDisconnectedPeers();
-	disconnectedPeers.data = disconnectedPeers
-		.map(orgPeer => refactorPeer(orgPeer, peerStates.DISCONNECTED));
+const getDisconnectedPeers = async params => {
+	const response = await getPeers(Object.assign(params, { state: 'disconnected' }));
+	return response;
+};
 
-	const data = [
-		...connectedPeers.data,
-		...disconnectedPeers.data,
-	];
-	const meta = { count: data.length };
-	return { data, meta };
+const getPeersStatistics = async () => {
+	const response = await peerCache.getStatistics();
+
+	return {
+		data: response,
+		meta: {},
+	};
 };
 
 module.exports = {
-	peerStates,
 	getPeers,
+	getConnectedPeers,
+	getDisconnectedPeers,
+	getPeersStatistics,
 };
